Remove warning if platform doesn't support CPU binding
Do not issue a warning if the platform doesn't support CPU binding
and the runtime tries to bind a thread to the set of accessible PUs (e.g.,
the progress thread in the communication layer). qthreads does not issue
a warning if threads can't be bound, so neither should we.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>