Fenced operations in message-order-fence mode must be delivery-complete
Fenced operations in message-order-fence MCM mode in COMM=ofi are used to
force visibility of previous operations. Previously, we did not also specify
FI_DELIVERY_COMPLETE, and as a result the fence only forced previous
operations to be transmit-complete, meaning they were transmitted but might
not be visible in memory. Specifying FI_DELIVERY_COMPLETE along with FI_FENCE
causes the previous operations to be visible.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>