Fix fillRandom wrong overload being selected
For filling random values for real arrays within a specified min and max
bound, the overload which results in a compiler error was being
selected.

For example:

```chapel
use Random;

var a : [0..<10] real;
fillRandom(a, min=-1, max=1);
```

would give us the following error:

```console
random.chpl:4: error: 'fillRandom' does not support non-rectangular arrays
```

Weirdly enough this same error does not appear when using an array of
`int`s instead of `real`s. This probably means there is a bug somewhere
with our where clauses. I say this because the isNumericOrBoolType() and
arr.isRectangular() checks both return true for the array `a` above but
still the error version of the proc is selected on main today.

A larger discussion about this would be nice but to get myself un-stuck
I would like to slap on some `pragma "last resort"`s to the offending
overloads which seems to fix the problem.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>