Change 'ref' alpha/beta arguments to 'const ref'
These alpha/beta arguments are declared '[const] ref' because the
Fortran implementations of the routine pass by pointer / C headers
take pointers, but should have been 'const ref' arguments rather than
'ref' in order to permit constants to be passed in as the actual
arguments.

This fixes all of the alpha/beta arguments that were using 'ref' to
use 'const ref' instead and updates our test to pass in 'const'
values to lock the behavior in better.

TODO:

- [ ] It looks like there are alpha/beta values passed by default
  intent; should these be updated as well, presumably?
- [ ] What other logical-'in' scalar arguments need a similar update?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>