Support calling randomOptimize with additional args that won't be optimized
There's some duplication here, which feels unfortunate.  Some of it is
necessary, because you can't pass "none" as an argument to a function when it
isn't expecting an extra argument (regardless of if the function is an fcf or
not).  But some of it is the result of not knowing how to make the call with an
arbitrary number of arguments, like in the original case

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>