Progress the comm layer while waiting for an atomic variable
If the atomic variable is going to be set to the value we are waiting for
based on a non-blocking operation that we initiated, then we may need to
progress the comm layer to ensure the operation is transmitted. Otherwise
it may sit in the network and we will wait forever.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>