Refactored rmaPutFn_msgOrdFence to inject properly
Refactored rmaPutFn_msgOrdFence to only inject non-blocking PUTs, because with
a blocking PUT the thread is going to block until transmit-complete, so there
is no advantage (and probably a slight disadvantage) to injecting it. Note
that injection is performed by setting FI_INJECT in the flags passed to
fi_writemsg and not by calling fi_inject_write. The former allows the buffer
to be reused immediately, but will generate a subsequent completion event.
This will allow non-blocking PUTs to be implemented in the future by allowing
the calling process to wait on the returned chpl_comm_nb_handle_t for the PUT
to complete. fi_inject_write does not generate a completion event so waiting
is not possible.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>