Fix amPutDone
Fixed a few problems with amPutDone. This function does PUT of a "done" flag
indicating to the initiating locale that the active message is complete.

1. All PUTs must be forced into visibility on all nodes for all MCM modes
except delivery-complete. The write-after-write ordering guarantee doesn't
help us because it is only valid on a pair of endpoints. We want to ensure
that all PUTs done by the AM are visible before we notify the initiator that
the AM is done.

2. The PUT must be blocking because we need to progress the endpoint to ensure
it is transmitted properly. Otherwise, there is no guarantee that we will
perform a subsequent operation on the same endpoint that will have the
side-effect of progressing the endpoint. If there isn't a subsequent
operation the PUT may never be transmitted. Even if the endpoint is
progressed by a subsequent operation it could be a long time until that
happens and the AM initiator will be delayed waiting for the notification
that the AM is done.

3. Because the PUT is blocking there is no benefit to injecting it.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>