Co-locale support for Nvidia GPUs
Support for partitioning Nvidia GPUs among co-locales. Each co-locale is
assigned the same number of GPUs, and the set of GPUs assigned to a
particular co-locale is based on the distance between the GPUs and the
co-locale's cores in the machine topology. In general, this means that
co-locales will use the GPUs that are "close" to it. Note that a greedy
algorithm is used to assign GPUs to co-locales, a result of which is that the
assignments may be sub-optimal especially for asymmetric machine topologies.
But on typical topologies it should do the right thing.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>