Improve the templates when the user doesn't have CHPL_HOME set
Michelle discovered that when a user installs using homebrew,
`$CHPL_HOME/util/printchplenv --anonymize` doesn't work because `$CHPL_HOME` is
not set.  List the alternative option Jade mentioned, though it does not
anonymize the output (being explicit that that is the case so it doesn't
surprise users).

Only the bug report and documentation report templates needed to be updated, the
others do not include the command in question

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>