Introduce 'sparseMatDat' type; reduce 'use's; rename SpsMatMatUtil->SpsMatUtil
The main thing here was to introduce a record called 'sparseMatDat'
that wraps and forwards to a map.  The goal here is to avoid having
the driver codes have to know about maps, and to support the ability
to substitute other implementations into the record over time, support
reductions on it, etc.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>