Touch up text around require statements w.r.t. -I / -L flags
[based on feedback and suggestions by @psath]

Based on feedback in #24963, this PR strives to improve the
documentation of the ``require`` statement to clarify that it doesn't
accept general `chpl` command-line flags and refactors the description
of the -I and -L flags to try and clarify their role in common to
listing files on the command-line or `require` statements.  While
here, I also did some wordsmithing and formatting adjustments.

Resolves #24963.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>