Clarify the issue templates w.r.t. workarounds and if the issue is blocking
We had a user recently seem confused whether calling the issue "blocking" was
talking about the state of the project or their personal experience.  In
response, it seemed like a good idea to adjust the request a bit.

Based on the structure of the feature request template, make the other two
templates talk about potential work-arounds in the "Description" section, and
make the request say "if any", so the user doesn't feel like they have to have
looked into work-arounds themselves.

Transform the question about if it is blocking to not mention work-arounds and
be more explicit about who is being blocked ("you", the person filing the
request, though I would expect Chapel developers filing such issues to use their
"user hat" and answer based on if they can accomplish what they are trying to
do when they encountered the issue)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>