Enable scope lookup to find containing module name
E.g. for this program

  module Program {
    module Helper {
      var debug = false;
    }
    var y = Program.Helper.debug;
  }

Previous to this change, y cannot be resolved, because it is unknown
what 'Program' refers to there. We have code like this in tests & these
compile with the production scope resolver. (This is a case where the
dyno scope resolver was leaving it unresolved).

Additionally, in the broader context of this PR, this is needed
for the test

  execflags/configs/ambiguousNestedModules.chpl

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>