Add notes about other ways to solve the Optimizers library argument problem
In talking to Jade and David during the user support/stabilization subteam
meeting today, I remembered a way Arkouda handles the "support multiple argument
types leading to genericness when I want to store things in the same data
structure" issue that this library will encounter when it extends the argument
types it supports.  Jade also reminded me of the existence of unions.  Note both
of these possible solutions for the future so we don't forget again by the time
we get back to this library

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>