Add mpf_get_str() and mp_exp_t
This adds a prototype of the mpf_get_str() routine and a test of
it.

This routine uses the 'mp_exp_t' type which we have not previously
defined in GMP.chpl, and which does not seem to have a well-defined
size in the GMP documentation.  As a result, I've defined it to be a
generic 'integral' type which seems to work in my runs, but may run
into valgrind issues (need to do more testing).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>