Set mp_exp_t to 'c_long' and add a check to make sure that's true
On linux64/chapcs, I was getting bogus output when passing in an int
value smaller than 'int(64)' for the 'expptr' argument.  This is not
completely surprising since we could be passing in a pointer to the
wrong-sized 'int' (and were) so could overwrite other important
memory.

Reviewing gmp.h, my sense is that this type is virtually always
'long int' so for now, I defined it to be 'c_long' and put in
an execution-time check at module initialization time to see
whether a 'sizeof()' check suggests we got the size wrong.

As a result of this, I also changed the argument intent for the
routine to 'out' which we can now do since it's no longer a generic
type, and because this more accurately reflects the semantics of
the argument (GMP doesn't care about the value on the way in).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>