Fix a sporadic output problem in the python interoperability begin test
Though rare, we were able to occasionally see the first `begin` call execute
after a later one.  The order of their output is not important, just that the
call completed at all before the end of the program (which it should).  With
that in mind, split the portions that are vulnerable to showing they're in a
different order to their own test files.

Technically, they could have just been moved into a single test file instead of
each to a separate one, but it didn't seem important one way or the other.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>