Add checks to gpuSort for calls from wrong locale
Resolves #24806

A user had requested better error messages when `gpuSort` is called on
an array which lives outside of the GPU.
This adds checks to make sure that gpuSort is only called on an array
from the GPU on which it lives. This error triggers even if gpuSort is
called from the wrong GPU (ex gpu1 when the array is on gpu0)

Once we have distributed arrays for GPUs, we will want to come back and
revisit this check.

Adds tests as well

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>