Fix spurious warning with misleading indentation
This was caused by the incorrect location reporting for variables.
If the code above is 4-indented, and then location of the name is used,
that name is also 4-indented, which flags as misleading.

The solution is to steal the logic from IncorrectIndentation that avoids
AST nodes that report their locations oddly.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>