Set CHPL_RT_LOCALES_PER_NODE even if it's 1
Previously, -nl Mx1 would not set CHPL_RT_LOCALES_PER_NODE, which was intended
to mimic the behavior before co-locales were implemented. However, if the
user had CHPL_RT_LOCALES_PER_NODE set to something other than 1 then there
would be an inconsistency between the launcher and the runtime. By setting it
to 1 both the launcher and runtime agree on the number of co-locales.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>