Enable remote sparse domain localization the quick and dirty way
This uses a new doi routine in a specific place to enable localization
of sparse domains when doing my new optimization.  Engin has told me
the more principled way to do it, but I wanted to get this into a commit
in a standalone branch first before refactoring.

Interestingly, my test here seems to show that we don't correctly
store 'definedConst' for sparse subdomains at present; though it
does work for const arrays, which is the case I actually care about
at present.  So future work here to generalize this.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>