Add Engin-requested test, assigns btwn arrays with same sparsity, diff domain
This adds a test Engin requested in his review to make sure that the
bulk assignment logic between sparse arrays is comparing domain index
sets precisely, and not relying on identity.  Though other tests
(including my not-yet-checked in sparse matrix-matrix multiplication)
arguably do this by virtue of the domain localization optimization,
this does it more expicitly by creating two domains with different
identities but the same indices, and assigning arrays back and forth
between them, modifying the arrays slightly along the way.

For a multilocale run, I'm creating the arrays on different locales,
but the domains on the same locale, somewhat arbitrarily; this
arguably covers a slightly different case than the existing
tests of the localization optimization, since in that case, the
domains and arrays are on different locales.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>