Update test to avoid memory leak
This is a longstanding test, but it isn't one that worked (ran to
completion) until I updated it to use a new optimization yesterday.
It has leaked memory all along, but because it didn't complete,
we didn't complain about it; now that it does, we started to.

Here, I'm changing the class types to use 'shared' (similar to what
was already done in its sibling, localBlock6.chpl) to close the leak
in a simple way without changing the point of the test (to see how
local blocks and classes interact).

Note that neither of these tests used any memory management styles
when originally written and that they were added later as we were
changing the way classes were managed by default.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>