Clarify how to call the module initialization function in interop scenarios
A user was confused about how to call the module initialization function when
using their Chapel library in another language.  While currently the function
takes two arguments, these arguments were not documented and are not needed in
the Fortran example that demonstrates the call.

Add documentation to the C section talking about these arguments but making it
clear that they may be removed in the future.  Be explicit in the Fortran
section that the arguments are not needed there, so that users are aware of the
difference in that scenario.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>