insertReturnTemps: ignore unresolved functions
For certain cases with parenless procs, call->resolvedOrVirtualFunction
can return a function that isn't resolved yet & has return type
dtUnknown. We don't know yet that it needs a return temp in that case &
it's equivalent to an unresolved Call (which must be handled somewhere
else). So, this change arranges for insertReturnTemps to calls
functions in that situation.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>