Add a no-local.good file for this test
I hadn't thought about how this test would behave in a no-local
setting, and am not sure I would've guessed the behavior would
match the multilocale .good file even if I had.  This adds
a no-local.good file to lock in the current behavior.

While here, I also noticed that I accidentally changed the
problem size of the domain without also changing the bbox in
the blockDist argument, and didn't mean for them not to match;
it doesn't seem to impact how the test works, but seemed worth
updating to avoid suggesting I was trying to do anything clever
or unusual—I wasn't.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>