Revert "gasnet/Makefile: --enable-pshm for all conduits"
This reverts commit 9a2d0ba30fcab88241ab957276fa3dd50a9fe064.

Enabling PSHM causes a startup performance regression when
`CHPL_GASNET_SEGMENT=fast` and perhaps other runtime performance regressions
as well. The former can be observed via the
test/performance/elliot/no-op.chpl test, which is about 9x slower on one
platform. The hypothesis is that this because of differences in the way that
the fixed heap is allocated with PSHM enabled vs. disabled. When PSHM is
disabled the heap is mmapped, and it's thought that the kernel is mapping all
pages at that time, so that chpl_comm_regMemHeapTouch isn't doing
anything. When PSHM is enabled the heap is not mapped, but
chpl_comm_regMemHeapTouch is only touching every 2MB so with a much smaller
page size there are many pages to be faulted in either when the heap is
registered or while the application runs. Either way, revert this change so
that PSHM is disabled by default and performance is restored.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>