Add an initialzer to LayoutCS
This adds an initializer to 'LayoutCS', taking the two bools that
characterize it.  This is motivated by my new compressRows-cs.chpl in
which I did not use named arguments, assumed that 'compressRows' was
the first argument (it is the first field in the 'CS' class, and ended
up (a) not testing what I intended to and (b) introducing a mysterious
memory leak, due to setting some bool other than the one I thought I
was.  While name-based argument passing would be the
obvious/safe/self-documenting thing to do (and the thing all our
existing tests of CS that pass arguments happen to do), it also seems
like we don't really want users setting fields in the abstract sparse
base classes.  So this PR introduces a new initializer taking the two
param bools saying whether to compress the rows or columns and whether
to store the indices in sorted order or not.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>