Restore field defaults
In my first draft here, I'd remoed the field defaults to remove
redundancy, but didn't think about the impact on existing code,
presumably due to being a fully-defualted generic type vs. not
(I'd assumed that the base classes would be sufficiently generic
that this wouldn't have an impact, but it did, to the tune of
a dozen or two failures.  Here, I'm restoring the field defaults
which keeps things working while still adding the new initializer.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>