Address bug 4723 in which ssh-spawner chokes tcsh
In bug 4723 it has been observed that newlines in environment
variables named in a `-E VAR` argument to ssh-spawner lead to errors
from `tcsh` when it is the user's login shell.  Follow-up tests show
that non-conforming variable names used by `bash` to export shell
functions are also an independent source of `tcsh` syntax errors.

Lacking any known (to us) portable means to quote a newline in the
`env VAR='VAL'` command used to implement `-E`, this commit instead
performs some filtering in the construction of the `env` command to
prevent the reported errors from arising.

 1. Omit propagation of any value containing a newline.
    This addresses the reported `Unmatched '.` errors from `tcsh`.
 2. Omit propagation of any variable with a non-conforming name.
    This addresses `Badly placed ()'s.` errors which occur when
    `tcsh` is parsing something like `env BASH_FUNC_module()=[...]`
    even when the `[...]` does not contain any newlines.

Note that this "addresses" the issue reported in Bug 4723, but does
*not* "resolve" it.  IMO, true resolution would support propagation of
environment variables (with conforming names) which contain newlines
in their value.

Cherry-picked from: gasnet:76c8a6a39

Signed-off-by: bonachea <dobonachea@lbl.gov>