Remove NUMA locale model from the docs
Since the `numa` locale model has been deprecated for a while, we should remove it
from the docs.
There are other docs which mention numa domains using the `-nl NxLt` syntax, which
still works and is unstable today so we keep those.

There are also other mentions of NUMA locale model in rst/developer/chips/inactive/18.rst
but since it is inactive, I left it as is.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>