third-party/gasnet/Makefile: Remove duplicate expansion of CHPL_GASNET_CFG_OPTIONS
Removes a harmless duplicate expansion of `$(CHPL_GASNET_CFG_OPTIONS)` in the
GASNet configure arguments.  GASNet configure arguments are all idempotent, so
passing the same arguments twice does not change behavior, it just
makes the configure line confusingly (much) longer for no reason. This, in turn,
degrades the human readability of the GASNet configure line, which can be
audited in the build log and GASNet's embedded ident strings.

Signed-off-by: Dan Bonachea <dobonachea@lbl.gov>