Add line separators to the list to avoid future mistakes
Both Jade and I find it very easy to mess up the formatting for lists.  While
the previous commit had a list that was well formatted, it'd be very easy for
future modifications to break it in ways that we'd only notice by looking at
the generated docs.  This commit makes that less likely, as the line spaces
should encourage future edits towards a style that definitely doesn't trigger
the formatting issue.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>