Specify that it's okay for formal names to be PascalCase as well
This was in the original wording but I wasn't sure why it was there because I
didn't recall any discussion about it.  But there also wasn't discussion that
they had to be camelCase only, either, just that we don't want snake_case at
all.  So restore the wording allowing PascalCase to those names

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>