Fix bug with returning the result of ternary expressions
Damian noticed that returning a string via a ternary expression resulted in
stack corruption.  It turns out that this was because the compiler pass
callDestructors is fragile - it is incomplete in its detection of variables that
don't need to be cleaned up at the end of their scope, and this incompleteness
resulted in inserting clean up for a variable that was still needed.

The pass itself makes a number of assumptions that are incorrect.  In
particular, it assumes:
- we can detect variables used to populate return variables by walking the AST
  backwards from the provided refStmt argument (we can only do that if refStmt
  is the last statement in the body)
- that each variable is only assigned into once (not true when the assignment
  occurs in a conditional)
- that there will only be one variable that we need to skip inserting
  deinitialization for in a particular scope (there could potentially be
  multiple)

We expect that implementing this pass correctly will be much easier in dyno.
Michael already has a start at it.  Therefore, it is okay to do a quick fix for
now, as long as we remember just how much is wrong with the pass today.

The quick fix in question is avoiding the insertion of cleanup for temporaries
marked with the flag `FLAG_IF_EXPR_RESULT`.  This allows the test in question to
pass without breaking the test cases that already worked.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>