turn concurrentMap test futures into nightly tests
Previously, we were locking down the expected memleak behavior for these
tests. Not all of our nightly jobs run futures but we would still like
to run these as correctness tests.

We also have infrastructure for detecting tests that leak memory so
it's better to leverage that rather than having these as futures
anyway.

So, I've rewritten these to not pass `--memLeaks` and then added some
separate "memLeak" only tests.

---
Signed-off-by: Andy Stone <stonea@users.noreply.github.com>