Have subprocess.sendPosixSignal throw IllegalArgumentError for bad signals
Before, this would throw a SystemError because `kill()` returns `EINVAL` for
bad signals and `EINVAL` does not have a SystemError subclass.  But since the
situations `kill()` throws `EINVAL` for are limited to bad signals (at least at
the time I looked it up), it seems fair to use `IllegalArgumentError` instead.
This also will allow users to distinguish between `EINVAL` that happened because
something went wrong setting up the subprocess and passing a bad signal.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>