chpl_run_utility1K should return 0 if no bytes are read
chpl_run_utility1K is documented to return the number of bytes read, but
returned -1 if zero bytes were read. Now it returns zero in that case. It's
unclear if it should print a warning in that situation. It seems to me that
the warning should be produced by the caller if reading zero bytes is
considered an error, but that's too substantial of a change for now.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>