Restore the recordio test to the proper place, d'oh!
I had misremembered and thought recordio was originally in the primers directory
but it turns out it was originally from the patterns directory.  That would
explain why there wasn't any sign of it in the history of the primers index file
for our online documentation.  Move the test to the right spot and remove it
from the wrong index file

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>