Updates based on feedback from Michael
Call out the errors that can be deferred from the `spawn` and `spawnshell` calls
there instead of in the documentation for the methods that might encounter them
as a result.  Have the methods impacted in that way reference that documentation
instead.

Also remove the error documentation from abort, alarm, kill, and terminate in
favor of just having the reference to sendPosixSignal's documentation.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>