Extend the ChplConfig and compiler flags test
Since I added a new flag and the new module-level variable, I wanted to make
sure they were tested and this test was pointed out as a good place to do that.

Because CHPL_LIB_PIC is not printed out without `--all`, I needed to adjust the
use of printchplenv to use that flag in order to verify it worked.  Doing so
pointed out that there were other environment variables that were in ChplConfig
but were not tracked by the test yet, so I added them to it while I was there.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>