Include interoperability commands when compiling with `--print-commands`
Prior to this change, `--library-python` compilation with `--print-commands`
would not include the Cython command used to make the Python library.  Including
the Cython command (and other helpful commands along the way) would allow faster
debugging and experimentation when the Cython command fails.

In more details, library compilation was using `runCommand` instead of the usual
`mysystem` commands that checked for `printSystemCommands`, due to wanting a
string result.  This commit made `runCommand` also check that boolean.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>