Add pointer in the man page to documentation on values for environment variables
Anna pointed out in #25514 that it was difficult to know what values were valid
for the new flag I added, but also that this was a general problem with the
flags in the compiler configuration options section.  The `--help` output seemed
generally more barebones, so didn't seem like a good place to add additional
information.  But the man page tends to have more information, so it seemed
reasonable to refer to the online documentation for the general category.

Note that this will not help if someone jumps to the flag they care about, but
it seemed annoying and repetitive to add the link to every single flag in the
section.  I'm open to suggestions on how to draw attention to the section
header, though

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>