Improve test cleanup for these tests
These tests were cleaning up old files, but not cleaning up the directory they
lived in.  Which is fine, but it led to errors in their prediff along the lines
of "Cannot create directory 'build': File exists" when the tests were re-run.
These did not cause the test to fail or anything, but it seemed like a good idea
to stop putting the tests in a situation where their prediff would do that.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>