Add more flag combination checks for the new chpldoc error message + test update
The test now covers explicitly setting `--output-dir` to docs when
`--save-sphinx` is used, and explicitly setting both to a different directory
that is still the same for the two flags.  Uses a new .good file for it.

Updates the .gitignore and .cleanfiles for the test, though we don't expect
those folders to be created now.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>