Add another flag variation, this time one that is expected to work
Validates that using "docs" for `--save-sphinx` is fine if `--output-dir` is
set to something else entirely.

This adds to the other .good files because I wanted to include a .catfiles to
grab the output for the version that works, and ensure we don't put things in
the other directories now that we error during chpldoc code instead of the
Sphinx stage.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>