Improve RangeChunk's handling of non-int(64) indices; fixes scans of same
This changes a loop in the RangeChunk module from using 0..#n to 0..<n.

The former results in an int(64) range while the latter results in a
n.type range since the former defines an unbounded int(64) range which
then determines the index type of any `#` operators applied to it
while the latter defines a range whose idxType is determined by both
its bounds, which are 0 and n.  This better preserves the idxType that
the RangeChunk operator is trying to compute with.

Since this code was used in our scan implementation, it broke scans
of non-int(64) idxType, and this improves the situation.

Resolves #25559.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>