Fix a bug with default values in Python interoperability
It turned out that the issue was because string arguments were causing us to
make a wrapper function and that wrapper function was what was getting exported.
Our default value tracking was assuming the argument stored wasn't going to get
replaced with another version after function resolution.

To fix, check for any formals that have a default value tracked in the old
function and insert a map entry for the new version of that function

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>