Implement idea from Jade to avoid re-exporting the entire Sort module
Makes a hidden internal module that does the re-export of the deprecated symbol,
and stifle the deprecation warning on that module.  This avoids deprecation
warnings the user can't do anything about while still generating them when they
would be relevant.  It also avoids re-exporting the rest of the Sort module,
which is much larger than we'd like.

I tested a version of this commit that only applied to the Search module in
another branch and the only fallout seemed to be due to having new modules
existing.  Will update the tests in the next commit

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>