Deprecate defaultComparator.compare and reverseComparator.compare argument names
`a` and `b` argument names do not match the style of other standard modules, so
replace them with `x` and `y`.  This should not have much impact for user code,
as the compare method is mostly used by the sort implementation itself and the
number and position of arguments is the same.

Additionally, specify that the type of the y argument is expected to be the same
as that of the x argument

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>