Create uAST to represent a type constructor
typeConstructorInitialQuery now builds a small module to store a
uAST Function that serves as the type constructor for a given type. This
module pretends to be an included module living under the module in
which the type was declared. For example, if a type 'R' was defined in
module 'foo', the included module would have the symbol path of
``foo.chpl__internal_foo_R``.

This module also includes a 'use' statement for the original module,
which is necessary in the event that fields' type/init expressions refer
to other symbols defined in that module.

Once finished building the uAST, we store the BuilderResult based on the
module path.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>