Enable chpl_comm_regMemHeapTouch() for more configurations
This commit updates the preprocessor logic which enables use of
`chpl_comm_regMemHeapTouch()` to more accurately reflect the set of
"configurations that register a fixed heap at startup".

Specifically, for registration-based conduits other than ibv, the code
for LARGE and FAST segment modes are identical. So, LARGE should be
excluded *only* for ibv-conduit, not for aries, ofi or ucx.

Signed-off-by: Paul H. Hargrove <PHHargrove@lbl.gov>