Use syntax-only mode for computing type genericity
This removes the need for the query system hack. The new code is adjusted
to still use substitution types when available, so that a trivial
R(string) is not considered generic because it has a now-instantiated
type parameter.

Also adjusts test12 in testTypePropertyPrimitives to use explicit generic form
(lest syntactic genericity check returns something weird).

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>